Form View Addition
Added Form View php file and added Form View button to the Search Table.  This is a basic example of how it will look once it's finished.  I'm thinking about having the button send the Row info to the Form View php file; but haven't much luck yet.  This is a visual example for the client meeting.