Geocode is retrieved from Report Form and saved in DB
If users fill in at least street, city, and zip the geocode function will go check for the lat and lng. Lat and lng are saved in DB upon submission.
Changed the regex pattern for street because it won't allow streets with two names (Sun Center Dr). Now it'll just  accept anything that is basically numbers followed by words.