Implemented Administrator Page (#114)
Based it off of volunteerlist.php. Has drop down for changing Volunteer Status. User level in SacFeralsUsers table will change according to Volunteer Status. If admin deletes record from VolunteerForm that has the same email in SacFeralsUsers, then that record will also be deleted. Removed form view and added reset for resetting user's passwords if account exists. Changed the regex in createaccount because there seems to be an issue with it.