Added vacation to dropdown and fix update bug in tables (#118)
Added vacation status for Volunteer Status. Added datestamp for when Admin change Volunteer Status to active/triage for the first time. I notice if you select certain columns in the table and try to update it with data the is the same as another row it won't update. Example: select only status column and change two of the rows to have the same status. The second column won't update because another row like it exists. I removed that logic from search.php and adminprofile.php. Correct me if I'm wrong about that logic.