Bug reports table not updating and lat/lng not showing up on record/form view (#166)
* Fixed record view update and Lat/Lng display on record/form view

Extra 's' in the update's prepare was preventing updating of DB. The $myArray[] stuff taken out caused the Lat/Lng to not show up properly. Added in some of the values taken out from $row to fix Lat/Lng too.  There was similar issues in form view that was taken care of.

Note: the issue with a comment field not being editable seem to be caused by my test values with html in it. For example in a comment field I put <table.....> which interfered with the actual html of the table.

* Undo changes in .iml