Need to specify NULL for autogenerated columns such as Id
Yeah there were still issues with our site on the new servers even if it is running PHP 5.6. I think the school might have several versions installed which is why we can get away with a lot of thing...idk. So I'm just going to try to fix whatever comes our way in terms of saving to DB.